chore: move all the hooks to hooks
